refactor: remove duplicate code and separate mock in another file
